requirements.txt
# Python项目依赖包
# 使用以下命令安装依赖：
# pip install -r requirements.txt

# 基础包
requests>=2.31.0
lxml>=4.9.0
beautifulsoup4>=4.12.0

# 如果需要处理XML的其他包
xmltodict>=0.13.0

# JSON to YAML Converter Dependencies
PyYAML>=6.0

# LangChain and OpenAI Dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
